+++ src/dipdup/datasources/datasource.py
@@ -17,6 +17,7 @@ from dipdup.http import HTTPGateway
 from dipdup.models import BigMapData
 from dipdup.models import HeadBlockData
 from dipdup.models import OperationData
+from dipdup.models import TokenTransferData
 from dipdup.utils import FormattedLogger
 
 _logger = logging.getLogger('dipdup.datasource')
@@ -25,6 +26,7 @@ _logger = logging.getLogger('dipdup.datasource')
 EmptyCallbackT = Callable[[], Awaitable[None]]
 HeadCallbackT = Callable[['IndexDatasource', HeadBlockData], Awaitable[None]]
 OperationsCallbackT = Callable[['IndexDatasource', Tuple[OperationData, ...]], Awaitable[None]]
+TokenTransfersCallbackT = Callable[['IndexDatasource', Tuple[TokenTransferData, ...]], Awaitable[None]]
 BigMapsCallbackT = Callable[['IndexDatasource', Tuple[BigMapData, ...]], Awaitable[None]]
 RollbackCallbackT = Callable[['IndexDatasource', MessageType, int, int], Awaitable[None]]
 
@@ -33,6 +35,8 @@ class Datasource(HTTPGateway):
     def __init__(self, url: str, http_config: HTTPConfig) -> None:
         super().__init__(url, http_config)
         self._logger = _logger
+        if http_config.cache:
+            self._logger.warning('`http.cache` option is deprecated. Implement caching logic manually if needed.')
 
     @abstractmethod
     async def run(self) -> None:
@@ -44,7 +48,6 @@ class Datasource(HTTPGateway):
 
 class HttpDatasource(Datasource):
     _default_http_config = HTTPConfig(
-        cache=True,
         retry_count=5,
         retry_sleep=1,
         ratelimit_rate=0,
@@ -74,6 +77,7 @@ class IndexDatasource(Datasource):
         self._on_disconnected_callbacks: Set[EmptyCallbackT] = set()
         self._on_head_callbacks: Set[HeadCallbackT] = set()
         self._on_operations_callbacks: Set[OperationsCallbackT] = set()
+        self._on_token_transfers_callbacks: Set[TokenTransfersCallbackT] = set()
         self._on_big_maps_callbacks: Set[BigMapsCallbackT] = set()
         self._on_rollback_callbacks: Set[RollbackCallbackT] = set()
         self._subscriptions: SubscriptionManager = SubscriptionManager(merge_subscriptions)
@@ -100,6 +104,9 @@ class IndexDatasource(Datasource):
     def call_on_operations(self, fn: OperationsCallbackT) -> None:
         self._on_operations_callbacks.add(fn)
 
+    def call_on_token_transfers(self, fn: TokenTransfersCallbackT) -> None:
+        self._on_token_transfers_callbacks.add(fn)
+
     def call_on_big_maps(self, fn: BigMapsCallbackT) -> None:
         self._on_big_maps_callbacks.add(fn)
 
@@ -119,6 +126,10 @@ class IndexDatasource(Datasource):
     async def emit_operations(self, operations: Tuple[OperationData, ...]) -> None:
         for fn in self._on_operations_callbacks:
             await fn(self, operations)
+
+    async def emit_token_transfers(self, token_transfers: Tuple[TokenTransferData, ...]) -> None:
+        for fn in self._on_token_transfers_callbacks:
+            await fn(self, token_transfers)
 
     async def emit_big_maps(self, big_maps: Tuple[BigMapData, ...]) -> None:
         for fn in self._on_big_maps_callbacks: